refactor(desktop): organize v2 CodeEditor into folder-per-module + dedupe view resolution
Per AGENTS.md "one folder per module with barrel index.ts":
- Extract inline helpers from CodeEditor.tsx into sibling extension folders:
  extensions/foldChevron, foldPlaceholder, contourSelectionLayer,
  selectionClassTogglePlugin. Hoists TRAILING_PAD / EMPTY_LINE_WIDTH_RATIO to
  module scope alongside their layer.
- Move loose CodeMirror utility modules into their own folders:
  CodeEditorAdapter/, createCodeMirrorTheme/, loadLanguageSupport/ (with
  streamLanguages co-located as its only consumer), syntax-highlighting/.
- Add registry/resolveActivePaneView/ — shared helper consumed by FilePane
  and FilePaneHeaderExtras to compute views + active view identically. Kills
  ~12 lines of duplication and removes one class of desync bugs.

CodeEditor.tsx drops from 425 → 256 lines, focused purely on the React component.
