+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/FilePane/registry/index.ts
@@ -1,5 +1,9 @@
 export { ALL_VIEWS } from "./allViews";
 export {
+	type ActivePaneView,
+	resolveActivePaneView,
+} from "./resolveActivePaneView";
+export {
 	orderForToggle,
 	pickDefaultView,
 	resolveViews,
