+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/FilePane/registry/resolveActivePaneView/index.ts
@@ -1,0 +1,4 @@
+export {
+	type ActivePaneView,
+	resolveActivePaneView,
+} from "./resolveActivePaneView";
