+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/FilePane/registry/views/CodeView/components/CodeEditor/CodeEditor.tsx
@@ -12,7 +12,7 @@ import {
 	indentOnInput,
 } from "@codemirror/language";
 import { highlightSelectionMatches, searchKeymap } from "@codemirror/search";
-import { Compartment, EditorSelection, EditorState } from "@codemirror/state";
+import { Compartment, EditorState } from "@codemirror/state";
 import {
 	drawSelection,
 	dropCursor,
@@ -21,12 +21,7 @@ import {
 	highlightActiveLineGutter,
 	highlightSpecialChars,
 	keymap,
-	type LayerMarker,
-	layer,
 	lineNumbers,
-	RectangleMarker,
-	ViewPlugin,
-	type ViewUpdate,
 } from "@codemirror/view";
 import { colorPicker } from "@replit/codemirror-css-color-picker";
 import { cn } from "@superset/ui/utils";
@@ -39,6 +34,10 @@ import {
 	createCodeMirrorAdapter,
 } from "./CodeEditorAdapter";
 import { createCodeMirrorTheme } from "./createCodeMirrorTheme";
+import { contourSelectionLayer } from "./extensions/contourSelectionLayer";
+import { buildFoldChevron } from "./extensions/foldChevron";
+import { buildFoldPlaceholder } from "./extensions/foldPlaceholder";
+import { selectionClassTogglePlugin } from "./extensions/selectionClassTogglePlugin";
 import { loadLanguageSupport } from "./loadLanguageSupport";
 import { getCodeSyntaxHighlighting } from "./syntax-highlighting";
 
@@ -51,152 +50,6 @@ interface CodeEditorProps {
 	editorRef?: MutableRefObject<CodeEditorAdapter | null>;
 	onChange?: (value: string) => void;
 	onSave?: () => void;
-}
-
-// Lucide chevron paths, inlined so we return a plain HTMLElement (foldGutter's
-// markerDOM contract) without bridging React. Matches lucide-react's ChevronDown
-// and ChevronRight exactly.
-const CHEVRON_DOWN_PATH = "m6 9 6 6 6-6";
-const CHEVRON_RIGHT_PATH = "m9 18 6-6-6-6";
-
-function buildFoldChevron(open: boolean): HTMLElement {
-	const el = document.createElementNS("http://www.w3.org/2000/svg", "svg");
-	el.setAttribute("xmlns", "http://www.w3.org/2000/svg");
-	el.setAttribute("viewBox", "0 0 24 24");
-	el.setAttribute("fill", "none");
-	el.setAttribute("stroke", "currentColor");
-	el.setAttribute("stroke-width", "2");
-	el.setAttribute("stroke-linecap", "round");
-	el.setAttribute("stroke-linejoin", "round");
-	el.setAttribute("class", "cm-foldChevron");
-	const path = document.createElementNS("http://www.w3.org/2000/svg", "path");
-	path.setAttribute("d", open ? CHEVRON_DOWN_PATH : CHEVRON_RIGHT_PATH);
-	el.appendChild(path);
-	return el as unknown as HTMLElement;
-}
-
-// Toggle a class on the editor root when any selection range is non-empty, so
-// CSS can suppress the active-line highlight while a selection is drawn.
-const selectionClassTogglePlugin = ViewPlugin.fromClass(
-	class {
-		constructor(view: EditorView) {
-			this.sync(view);
-		}
-		update(update: ViewUpdate) {
-			if (update.selectionSet || update.docChanged) {
-				this.sync(update.view);
-			}
-		}
-		sync(view: EditorView) {
-			const hasSelection = view.state.selection.ranges.some((r) => !r.empty);
-			view.dom.classList.toggle("cm-hasSelection", hasSelection);
-		}
-	},
-);
-
-// Custom selection layer: draws selection backgrounds per-line, snug to each
-// line's actual text instead of CM's default full-line-width fill for middle
-// lines of multi-line selections.
-//
-// We keep drawSelection() for cursor rendering (including multi-cursor); its
-// own .cm-selectionBackground rectangles are hidden via CSS so this layer is
-// the only thing painting selection backgrounds.
-const contourSelectionLayer = layer({
-	above: false,
-	class: "cm-contourSelectionLayer",
-	update(update) {
-		return (
-			update.docChanged ||
-			update.viewportChanged ||
-			update.selectionSet ||
-			update.geometryChanged
-		);
-	},
-	markers(view) {
-		const markers: LayerMarker[] = [];
-		const lineHeight = view.defaultLineHeight;
-		for (const range of view.state.selection.ranges) {
-			if (range.empty) continue;
-			const fromLine = view.state.doc.lineAt(range.from);
-			const toLine = view.state.doc.lineAt(range.to);
-			const TRAILING_PAD = 4;
-			// Half-height-wide stub for empty lines in the middle of a selection
-			// so the selection stays visually contiguous through blank lines.
-			const EMPTY_LINE_WIDTH = Math.round(lineHeight / 2);
-			for (let n = fromLine.number; n <= toLine.number; n += 1) {
-				const line = view.state.doc.line(n);
-				const selStart = Math.max(range.from, line.from);
-				// Clamp selection end to actual text end so trailing whitespace
-				// space past the last visible character is never filled.
-				const textEnd = line.from + line.text.length;
-				const selEnd = Math.min(range.to, textEnd);
-				const isEmpty = selStart >= selEnd;
-				const isMiddleLine = n > fromLine.number && n < toLine.number;
-				// Skip edge lines that fall in empty territory (selection starts at
-				// end-of-line or ends at start-of-line); only show the stub for
-				// genuinely empty middle lines.
-				if (isEmpty && !isMiddleLine) continue;
-				const lineRange = isEmpty
-					? EditorSelection.cursor(line.from)
-					: EditorSelection.range(selStart, selEnd);
-				for (const m of RectangleMarker.forRange(
-					view,
-					"cm-contourSelection",
-					lineRange,
-				)) {
-					// Expand each rect to fill the full line-cell height. Use exactly
-					// lineHeight (no +1) so consecutive rects abut without overlap —
-					// overlap darkens at transparent fill alphas into a visible stripe.
-					const gap = Math.max(0, lineHeight - m.height);
-					const width = isEmpty
-						? EMPTY_LINE_WIDTH
-						: (m.width ?? 0) + TRAILING_PAD;
-					markers.push(
-						new RectangleMarker(
-							"cm-contourSelection",
-							m.left,
-							m.top - gap / 2,
-							width,
-							lineHeight,
-						),
-					);
-				}
-			}
-		}
-		return markers;
-	},
-});
-
-// Lucide MoreHorizontal (three dots) — inline SVG built imperatively so we can
-// return a plain HTMLElement to CM's placeholderDOM contract.
-function buildFoldPlaceholder(
-	_view: unknown,
-	onclick: (event: Event) => void,
-): HTMLElement {
-	const button = document.createElement("button");
-	button.type = "button";
-	button.className = "cm-foldPlaceholder";
-	button.setAttribute("aria-label", "Unfold");
-	button.addEventListener("click", onclick);
-
-	const svg = document.createElementNS("http://www.w3.org/2000/svg", "svg");
-	svg.setAttribute("xmlns", "http://www.w3.org/2000/svg");
-	svg.setAttribute("viewBox", "0 0 24 24");
-	svg.setAttribute("fill", "none");
-	svg.setAttribute("stroke", "currentColor");
-	svg.setAttribute("stroke-width", "2");
-	svg.setAttribute("stroke-linecap", "round");
-	svg.setAttribute("stroke-linejoin", "round");
-	svg.setAttribute("class", "cm-foldPlaceholderIcon");
-	for (const cx of ["5", "12", "19"]) {
-		const c = document.createElementNS("http://www.w3.org/2000/svg", "circle");
-		c.setAttribute("cx", cx);
-		c.setAttribute("cy", "12");
-		c.setAttribute("r", "1");
-		svg.appendChild(c);
-	}
-	button.appendChild(svg);
-	return button;
 }
 
 export function CodeEditor({
@@ -257,14 +110,7 @@ export function CodeEditor({
 				highlightActiveLineGutter(),
 				highlightSpecialChars(),
 				history(),
-				// Render fold markers as Lucide SVGs rather than Unicode glyphs —
-				// text glyphs have font-dependent baselines that refuse to align
-				// with line numbers. SVGs have exact bounding boxes and scale
-				// predictably. Hover reveal is handled in createCodeMirrorTheme.
 				foldGutter({ markerDOM: buildFoldChevron }),
-				// Collapsed-block placeholder uses Lucide MoreHorizontal. Lives
-				// in a separate codeFolding() — its config facet combines with
-				// the one registered internally by foldGutter().
 				codeFolding({ placeholderDOM: buildFoldPlaceholder }),
 				drawSelection(),
 				dropCursor(),
