+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/FilePane/registry/views/CodeView/components/CodeEditor/CodeEditorAdapter/index.ts
@@ -1,0 +1,5 @@
+export {
+	type CodeEditorAdapter,
+	createCodeMirrorAdapter,
+	type EditorSelectionLines,
+} from "./CodeEditorAdapter";
