+++ apps/desktop/src/renderer/routes/_authenticated/_dashboard/v2-workspace/$workspaceId/hooks/usePaneRegistry/components/FilePane/registry/views/CodeView/components/CodeEditor/loadLanguageSupport/index.ts
@@ -1,0 +1,1 @@
+export { loadLanguageSupport } from "./loadLanguageSupport";
